requirements.txt
langchain==0.0.148
openai==0.28.0
python-dotenv==1.0.0